Metaoroif NFT
Metaroid is an NFT marketplace focussed on social features that attract user engagement. Our platform provides gamification and it opens space for metaverse owing to 1155 NFT format which enables the NFT covering single to multiple limited authentic exemplars and creator wallets. The issued licenses or certificates are able to grasp physical and digital assets